fix: enforce available balance across loan lifecycle

+++ caixinha-serverless/renegociar/index.js
@@ -4,6 +4,7 @@ const middleware = require('../utils/middleware')
 const { connect, getByIdOrThrow, insertDocument, deleteByProjection, replaceDocumentById } = require('../v2/mongo-operations')
 const dispatch = require('../amqp/events')
 const { resolveCircularStructureBSON } = require('../utils')
+const coreVersion = require('caixinha-core/package.json').version
 
 const COLLECTION_RENEGOCIACOES = 'renegociacoes'
 const COLLECTION_EMPRESTIMOS = 'emprestimos'
@@ -22,6 +23,8 @@ function setMemberInstance(renegociacaoJson) {
 
 async function atualizarRenegociacao(renegociacao, boxID, renegociacaoId) {
     renegociacao['boxId'] = boxID
+    renegociacao['coreVersion'] = coreVersion
+    renegociacao['balanceMutationApplied'] = false
     await replaceDocumentById(renegociacaoId, COLLECTION_RENEGOCIACOES, resolveBSONStructureRenegociacao(renegociacao))
 }
 
@@ -58,11 +61,13 @@ async function handle(context, req) {
     await connect()
 
     const renegociacao = await getByIdOrThrow(renegociacaoId, COLLECTION_RENEGOCIACOES)
-    const caixinha = Box.fromJson(await getByIdOrThrow(renegociacao.boxId))
+    const boxEntity = await getByIdOrThrow(renegociacao.boxId)
+    const caixinha = Box.fromJson(boxEntity)
     renegociacao['oldLoan']['box'] = caixinha
     setMemberInstance(renegociacao)
 
     const reneg = Renegotiation.fromJson(renegociacao)
+    const balanceBefore = caixinha.balance
 
     const { reneg: novaRenegociacao, newLoan } = AcceptRenegotiation(
         caixinha,
@@ -74,11 +79,19 @@ async function handle(context, req) {
         }
     )
 
+    if (Math.abs(caixinha.balance - balanceBefore) > 0.005) {
+        const error = new Error('Renegociacao alterou o saldo da caixinha')
+        error.code = 'RENEGOTIATION_BALANCE_INVARIANT_VIOLATION'
+        throw error
+    }
+
     const { deletedCount } = await deleteByProjection({ uid: novaRenegociacao.originLoan.UUID }, COLLECTION_EMPRESTIMOS)
     context.log('deleted count', deletedCount)
 
     await salvarNovoEmprestimo(newLoan)
-    await replaceDocumentById(renegociacao.boxId, 'caixinhas', resolveCircularStructureBSON(caixinha))
+    await replaceDocumentById(renegociacao.boxId, 'caixinhas', resolveCircularStructureBSON(caixinha), {
+        expectedVersion: boxEntity._version
+    })
     await atualizarRenegociacao(novaRenegociacao, renegociacao.boxId, renegociacaoId)
     notificar(reneg.owner, renegociacaoId, renegociacao.boxId)
 }
